refactor : move note to notes
